+++ include/qmckl_memory.h
@@ -16,7 +16,7 @@
 void *qmckl_malloc_host(qmckl_context_device context,
 						const qmckl_memory_info_struct_device info);
 
-void *qmckl_malloc_device(qmckl_context_device context, size_t size);
+void *qmckl_malloc_device(qmckl_context_device context, int64_t size);
 
 qmckl_exit_code_device qmckl_free_host(qmckl_context_device context,
 									   void *const ptr);
@@ -28,10 +28,10 @@ qmckl_exit_code_device qmckl_free_device(qmckl_context_device context,
 
 qmckl_exit_code_device qmckl_memcpy_H2D(qmckl_context_device context,
 										void *const dest, void *const src,
-										size_t size);
+										int64_t size);
 qmckl_exit_code_device qmckl_memcpy_D2H(qmckl_context_device context,
 										void *const dest, void *const src,
-										size_t size);
+										int64_t size);
 qmckl_exit_code_device qmckl_memcpy_D2D(qmckl_context_device context,
 										void *const dest, void *const src,
 										size_t size);
